refactor(kubernetes-deployment): split flux.go into phase-grouped files
The 875-line flux.go is split into five files inside the same module
(no breaking change to callers — all functions stay on KubernetesDeployment):

- flux.go            — package-private helpers (fluxCliContainer, parseTimeout)
- flux-render.go     — FluxRenderConfig, FluxCommitConfig
- flux-secrets.go    — ValidateAgeKeyPair, FluxEncryptSecrets, FluxApplySecrets, FluxVerifySecrets
- flux-apply.go      — FluxDeployOperator, FluxApplyConfig, FluxWaitForReconciliation
- flux-bootstrap.go  — FluxBootstrap, FluxDestroy (lifecycle orchestrators)

Function signatures and bodies are byte-identical to the original
(verified via per-function md5 of the signature block).

Refs #143 (P1, Option A — top-level `flux` module deferred until #3 is
tackled together).
